Move TestService and TestModule to separate files
A workaround for some 8u202 compiler bug
