+++ exonum-java-binding/core/src/test/java/com/exonum/binding/runtime/TestServiceModule.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2019 The Exonum Team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.exonum.binding.runtime;
+
+import com.exonum.binding.service.Service;
+import com.exonum.binding.service.TransactionConverter;
+import com.google.inject.AbstractModule;
+
+class TestServiceModule extends AbstractModule {
+
+  @Override
+  protected void configure() {
+    bind(Service.class)
+        .to(TestService.class);
+
+    bind(TransactionConverter.class)
+        .toInstance((m) -> (context) -> System.out.println("Transaction#execute"));
+  }
+}
